CB-17581 Extract environment consumption logic from handlers into a separate service
* Follow-up to https://jira.cloudera.com/browse/CB-17216 / #12807.
* Rename the current
  com.sequenceiq.environment.environment.service.consumption.
  ConsumptionService to ConsumptionClientService (keeping the package
  unchanged), so that the latter new service only encapsulates the web
  service calls, nothing else.
* Extract all the consumption-related logic from
  com.sequenceiq.environment.environment.flow.creation.handler
  .StorageConsumptionCollectionSchedulingHandler and
  com.sequenceiq.environment.environment.flow.deletion.handler
  .StorageConsumptionCollectionUnschedulingHandler into
  com.sequenceiq.environment.environment.service.consumption.
  ConsumptionService. (This is the same service mentioned in the first
  bullet item.)
* Testing:
  * Manual testing in local CB.
  * Added new UT and updated existing ones.
